finished implementing backend code for currency conversion
